requirements.txt
PySide6>=6.5.0
requests>=2.31.0
argon2-cffi>=23.1.0
python-nmap>=0.7.1
scapy>=2.5.0
dnspython>=2.4.0
python-docx>=1.1.0
pyinstaller>=6.4.0
fpdf2>=2.7.0
reportlab>=4.0.0
Jinja2>=3.1.2
matplotlib>=3.8.0
beautifulsoup4>=4.12.0
psycopg2-binary>=2.9.9
cryptography>=42.0.0
psutil>=5.9.8
PyYAML>=6.0.1

